refactor https://github.com/microsoft/vscode/pull/256491 to support existing policy generation infrastructure
